tests/test_models.py
import unittest
from myapp.models import User, db


class TestUserModel(unittest.TestCase):
    """
    Test cases for the User model.

    Attributes:
        db (SQLAlchemy): The SQLAlchemy database object used for testing.
    """

    def setUp(self):
        """
        Set up the test environment before each test.
        """
        db.create_all()

    def tearDown(self):
        """
        Clean up after each test if needed.
        """
        db.session.remove()
        db.drop_all()

    def test_user_creation(self):
        """
        Test creating a User object and adding it to the database.

        Checks:
            - User object creation.
            - Assignment of an ID to the User after creation.
            - Other relevant assertions can be added.
        """
        user = User(
            first_name='John',
            last_name='Doe',
            email='john@example.com',
            # Add other required fields
        )
        db.session.add(user)
        db.session.commit()

        self.assertIsNotNone(user.id)  # User has been assigned an ID after creation
        # Add more assertions to check field values and constraints

    def test_settings_creation(self):
        """
        Test creating a Settings object and adding it to the database.

        Checks:
            - Settings object creation.
            - Assignment of attributes to the Settings object after creation.
            - Other relevant assertions can be added.
        """
        settings = Settings(
            user_id=1,  # Replace with a valid user ID for testing
            template_mode='light'
            # Add other required fields
        )
        db.session.add(settings)
        db.session.commit()

        self.assertIsNotNone(settings.id)  # Settings has been assigned an ID after creation
        # Add more assertions to check field values and constraints

        # Add more test methods to cover various aspects of the User model

    def test_customer_creation(self):
        """
        Test creating a Customer object and adding it to the database.

        Checks:
            - Customer object creation.
            - Assignment of attributes to the Customer object after creation.
            - Other relevant assertions can be added.
        """
        customer = Customer(
            first_name='Alice',
            last_name='Smith',
            email='alice@example.com',
            phone_number='1234567890',
            shipping_address='123 Main St'
            # Add other required fields
        )
        db.session.add(customer)
        db.session.commit()

        self.assertIsNotNone(customer.id)  # Customer has been assigned an ID after creation
        # Add more assertions to check field values and constraints

    def test_transactions_creation(self):
        """
        Test creating a Transactions object and adding it to the database.

        Checks:
            - Transactions object creation.
            - Assignment of attributes to the Transactions object after creation.
            - Other relevant assertions can be added.
        """
        transaction = Transactions(
            product_name='Product A',
            product_description='Description of Product A',
            order_date='2023-11-23',
            delivery_address='123 Main St',
            delivery_date='2023-11-30',
            total_price=100.0,
            delivery_fee=10.0,
            invoice_link='http://example.com/invoice',
            discount_applied=5.0,
            amount_paid=90.0,
            remaining_balance=5.0,
            due_date='2023-12-05',
            payment_status='partial'
            # Add other required fields
        )
        db.session.add(transaction)
        db.session.commit()

        self.assertIsNotNone(transaction.id)  # Transactions has been assigned an ID after creation
        # Add more assertions to check field values and constraints


if __name__ == '__main__':
    unittest.main()
